changing to use the cmdio renderer
