+++ cmd/pipelines/show.go
@@ -13,7 +13,6 @@ import (
 	"github.com/databricks/cli/libs/databrickscfg/cfgpickers"
 	"github.com/databricks/cli/libs/flags"
 	"github.com/databricks/cli/libs/sqlexec"
-	"github.com/databricks/cli/libs/tableview"
 	"github.com/databricks/databricks-sdk-go"
 	"github.com/spf13/cobra"
 )
@@ -104,7 +103,7 @@ func render(ctx context.Context, cmd *cobra.Command, columns []string, rows [][]
 		if len(columns) == 0 {
 			return nil
 		}
-		return tableview.RenderStaticWithTruncation(ctx, out, columns, rows, tableview.DetectWidth(out))
+		return renderTable(ctx, columns, rows, detectWidth(out))
 	default:
 		return fmt.Errorf("unknown output type %s", root.OutputType(cmd))
 	}
